backend/requirements.txt
flask>=3.0
flask-cors>=4.0
torch>=2.1
torchvision>=0.16
numpy>=1.24
scipy>=1.11
scikit-image>=0.21
tifffile>=2023.7
cellpose>=4.0
faiss-cpu>=1.7.4
Pillow>=10.0
